Renomeia arquivos de classe conforme conteúdo
