+++ src/E5R.Architecture.Core/PaginatedResult`1.cs
